+++ lib/parser_engine/parser_engine.dart
@@ -1,6 +1,7 @@
 library parser_engine;
 
 part './vietnamese_parser_engine.dart';
+part './vietnamese_parser_engine_2.dart';
 
 abstract class ParserEngine {
   String unsigned(final String text);